pyproject.toml
[tool.poetry]
name = "bank-account-management"
version = "0.1.0"
description = ""
authors = ["Degisew <jegisew21@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
django-allauth = "^65.3.0"
model-bakery = "^1.20.0"
djangorestframework = "^3.15.2"
dj-rest-auth = "^7.0.0"
requests = "^2.32.3"
django-model-utils = "^5.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
coverage = "^7.6.9"
pytest-cov = "^6.0.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
